Add files via upload
I worked as Business Intelligence Associate at Shopee at the moment when I code this scripts.

Context: This was a request from the Business Developer Manager to retrieve more information regarding sellers, such as:
- Best sellers in MercadoLibre site (key fields: orders last 60 days, seller name and seller id)
- SKUS with the best sales (key fields: sku_id, sku_name, seller_id, price, sold_quantity)

In a month, I was able to scrape more than 19 million of SKU from the site in less than a day, this info was used to set goals to relational managers (i.e. RM, similar to Key Account Managers), so RM can hunt sellers and specially those SKU that have best sales in MELI, letting us to upload those SKU from the same seller to our platoform. In a coupe of weeks, we were able to know which SKUS we need to get in our playtform and the competitive price to improve sales.